+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 ### [Changed]
 * Change the theme to use `nerdtree_syntax_dark` for the plugin `chadtree`.
 * Replace `ruff_lsp` with `ruff` to format python code because `ruff_lsp` will be deprecated by neovim lspconfig.
+* Change shortcuts for the plugin `telescope.nvim`:
+	* `<leader>ff`: `<cmd>telescope find_files`
+	* `<leader>fg`: `<cmd>telescope live_grep`
+	* `<leader>fr`: `<cmd>telescope oldfiles`
 
 ### [Fixed]
 * Change the source of the plugin `mason.nvim` to use my fork because I has been added some code to fix the bug. The bug is that the plugin `mason.nvim` couldn't install some langauge servers if I use `pyenv-win` on windows.